Refactor client component structure, fix select bar z index, add titles to container bar settings
