+++ packages/client/src/components/preview/Indicator.svelte
